% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlGenScores.R
\name{getPlGenScores}
\alias{getPlGenScores}
\title{getPlGenScores}
\usage{
getPlGenScores(mat, plGen, exclude = NULL)
}
\arguments{
\item{mat}{Matrix of regional plant - bee interactions (i.e. potential interactions to draw from). Rows = plants, cols = bees.}

\item{plGen}{Character vector of current plant names from each location}

\item{exclude}{Character vector of plant names that should not be considered (e.g. noxious weed, non-native)}
}
\value{
Data frame with the following columns:
\itemize{
\item PlantGenus: suggested genus of plant to add
\item N: new bee species that would be added
\item Chao1: (Chao1) visitor richness for this plant, estimated from interaction matrix
\item ChaoZ: Z-score of difference in Chao1 scores (i.e. "Does this species significantly change richness?")
\item PropShannon: Proportion increased Shannon-Weiner diversity
}
}
\description{
Function to get "scores" for unlisted plant genera (i.e. how much would bee species counts/network properties be increased by adding new species?)
}
\examples{
set.seed(12)
#Regional interaction matrix
iMat <- matrix(rpois(100,lambda = rgamma(n = 100,.1,.1)),10,10,
  dimnames = list(paste0('plant',LETTERS[1:10]),paste0('bee',LETTERS[1:10])))
  
getPlGenScores(mat = iMat, #Regional network (plant spp - bee spp)
  plGen = c('plantA','plantB','plantC'), #Plants found at focal vineyard
  exclude = c('plantG','plantJ')) #Excluded plants

}
