% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGenSpp.R
\name{makeGenSpp}
\alias{makeGenSpp}
\title{Create genus/species column}
\usage{
makeGenSpp(d, genCol, sppCol, newCol = genSpp)
}
\arguments{
\item{d}{Data frame or tibble}

\item{genCol}{Genus column}

\item{sppCol}{Species column}

\item{newCol}{New column name (default = genSpp)}
}
\value{
Data frame with new genSpp column
}
\description{
Creates genus/species column, accounting for empty species labels (adds "spp."). Originally included in \emph{SRtools} package
}
\examples{
dat <- data.frame(gen=c('Apis','Bombus','Nomada'),spp=c('mellifera','nevadensis',NA))
makeGenSpp(dat,gen,spp)
}
