% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReports.R
\name{makeReports}
\alias{makeReports}
\title{Make vineyard reports}
\usage{
makeReports(
  plantListCSV = NA,
  beeDataCSV = NA,
  iNatFolder = NA,
  reportFolder = NA,
  vinePlDatCSV = NA,
  predictedBeesCSV = NA,
  dataStoragePath = NA,
  famGenPath = "./data/famGenLookup.csv",
  orCountyShpPath = "./data/shapefiles/orcntypoly.shp",
  orEcoregShpPath = "./data/shapefiles/or_eco_l3.shp",
  beeAbstractsPath = "./data/Bee_Abstracts.csv"
)
}
\arguments{
\item{plantListCSV}{CSV of plants from Oregon Flora (csv)}

\item{beeDataCSV}{CSV of bee/plant interactions from OBA}

\item{iNatFolder}{Folder containing vineyard iNaturalist CSV files. Looks in sub-folders as well}

\item{reportFolder}{Folder for writing reports to}

\item{vinePlDatCSV}{(Optional) output csv of all vineyard records. Skips writing if NULL}

\item{predictedBeesCSV}{(Optional) output csv of predicted bees for each vineyard. Skips writing if NULL}

\item{dataStoragePath}{(Optional) .Rdata storage path for internal function data. Skips writing if NULL}

\item{famGenPath}{(Optional) Path to bee genus-family lookup csv}

\item{orCountyShpPath}{(Optional) Path to Oregon county polygons}

\item{orEcoregShpPath}{(Optional) Path to Oregon ecoregion polyogons}

\item{beeAbstractsPath}{(Optional) Path to bee/plant Abstract csv}
}
\value{
Nothing - writes to vinePlDatCSV, predictedBeesCSV, or dataStoragePath
}
\description{
Create vineyard reports from iNaturalist project data, bee/plant interactions, and plant list.
}
